Conexión front y back.
Se conecta el componente Sobre-mi al backend.

+++ Back-End/src/main/java/com/portfolio/lau/Repository/PersonaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonaRepository extends JpaRepository <Persona, Long> {
+public interface PersonaRepository extends JpaRepository <Persona, Integer> {
     public Optional<Persona> findByNombre(String nombre);
     public boolean existsByNombre(String nombre);
 }
